# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DynamicsERP API specification — Solar Projects ERP
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: projectId, schema: { type: integer } }
        - { in: query, name: submittedById, schema: { type: integer } }
        - { in: query, name: category, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Expense" } } } } }
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateExpense" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateExpense" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
  /expenses/{id}/submit:
    post:
      operationId: submitExpense
      tags: [expenses]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
  /expenses/{id}/approve:
    post:
      operationId: approveExpense
      tags: [expenses]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
  /expenses/{id}/reject:
    post:
      operationId: rejectExpense
      tags: [expenses]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody: { required: false, content: { application/json: { schema: { $ref: "#/components/schemas/RejectAction" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
  /expenses/{id}/reimburse:
    post:
      operationId: reimburseExpense
      tags: [expenses]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }
...
components:
  schemas:
...
    RejectAction:
      type: object
      properties:
        reason: { type: ["string", "null"] }
...
    Expense:
      type: object
      properties:
        id: { type: integer }
        expenseNumber: { type: string }
        submittedById: { type: integer }
        submittedByName: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        category: { type: string }
        amount: { type: number }
        expenseDate: { type: string, format: date }
        description: { type: ["string", "null"] }
        status: { type: string }
        receiptUrl: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        approvedById: { type: ["integer", "null"] }
        approvedByName: { type: ["string", "null"] }
        approvedAt: { type: ["string", "null"], format: date-time }
        rejectedReason: { type: ["string", "null"] }
        reimbursedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, expenseNumber, submittedById, category, amount, expenseDate, status, createdAt, updatedAt]
    CreateExpense:
      type: object
      properties:
        projectId: { type: ["integer", "null"] }
        category: { type: string }
        amount: { type: number }
        expenseDate: { type: string, format: date }
        description: { type: ["string", "null"] }
        receiptUrl: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [category, amount, expenseDate]
    UpdateExpense:
      type: object
      properties:
        projectId: { type: ["integer", "null"] }
        category: { type: ["string", "null"] }
        amount: { type: ["number", "null"] }
        expenseDate: { type: ["string", "null"], format: date }
        description: { type: ["string", "null"] }
        receiptUrl: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...